Added triggered mode to the forearm camera driver, and added a timestamp to the outgoing image. The timestamp should be very accurate (1 ms) in the triggered case, and is usually equally accurate in the untriggered case.
So far the camera is occasionally triggering early, a condition which is
not currently well handled, so the timestamp will be incorrect when this
happens.

I had to modify the interface to the camera library slightly in order to
get as accurate timing information as posible. While I was at it a also
removed a few annoying warnings.



git-svn-id: https://svn.code.sf.net/p/personalrobots/code/pkg/trunk@13989 f5854215-dd47-0410-b2c4-cdd35faa7885

Former-commit-id: 174826922ad1244622de4c51168b124709711d7a

+++ drivers/cam/forearm_cam/src/pr2lib/host_netutil.c
@@ -442,7 +442,7 @@ int wgWaitForPacket( int s, uint32_t type, size_t pktLen, uint32_t *wait_us ) {
 			// All valid WG command packets have magic_no == WG_MAGIC NO
 			// We also know the minimum packet size we're looking for
 			// So we can drop short or invalid packets at this stage
-			if( (r < pktLen) ||
+			if( ((unsigned int) r < pktLen) ||
 						gPkt.magic_no != htonl(WG_MAGIC_NO) ||
 						gPkt.type != htonl(type) ) {
 				debug("Dropping packet with magic #%08X, type 0x%02X (looking for 0x%02X), length %d (looking for %d)\n", ntohl(gPkt.magic_no), ntohl(gPkt.type), type, r, pktLen);
